+++ src/plugins/icons/registry.ts
@@ -1,0 +1,6 @@
+import { featherProvider } from "./feather-provider";
+import { IconSetProvider } from "./types";
+
+export const providerRegistry = new Map<string, IconSetProvider>([
+  [featherProvider.name, featherProvider],
+]);